+++ R/io_readLAS.R
@@ -25,14 +25,16 @@
 #
 # ===============================================================================
 
-
-
 #' Read .las or .laz files
 #'
 #' Reads .las or .laz files into an object of class \link[=LAS-class]{LAS}. If several files are read at
 #' once the returned LAS object is considered as one LAS file. The optional parameters enable the user
 #' to save a substantial amount of memory by choosing to load only the attributes or points of interest.
-#' The LAS formats 1.1 to 1.4 are supported. Point Data Record Format 0,1,2,3,5,6,7,8 are supported.
+#' The LAS formats 1.1 to 1.4 are supported. Point Data Record Format 0,1,2,3,5,6,7,8 are supported.\cr\cr
+#' \code{readLAS} is the generic function. Using one of \code{read*LAS} functions adds a tag to the returned
+#' object to register a point cloud type. Registering the good point type \bold{may}
+#'  improve the performance of some functions (see also \link[=LAS-class]{LAS-class}.)
+#'
 #'
 #' \strong{Select:} the 'select' argument specifies the data that will actually be loaded. For example,
 #' 'xyzia' means that the x, y, and z coordinates, the intensity and the scan angle will be loaded.
@@ -49,7 +51,7 @@
 #' before reading the file, the internal filters should always be preferred. The available filters are
 #' those from \code{LASlib} and can be found by running the following command: \code{readLAS(filter = "-help")}.
 #' (see also \link[rlas:read.las]{rlas::read.las}). From \code{rlas} v1.3.6 the transformation commands
-#' can also be passed via the argument filter.
+#' can also be passed via the argument filter.\cr\cr
 #'
 #' @param files characters. Path(s) to one or several a file(s). Can also be a
 #' \link[=LAScatalog-class]{LAScatalog} object.