+++ R/normalize_height.R
@@ -198,6 +198,7 @@ normalize_height.LAS = function(las, algorithm, na.rm = FALSE, use_class = c(2L,
 
   fast_quantization(las@data[["Z"]], zscale, zoffset)
   las <- lasupdateheader(las)
+  las@type <- las@type + NLAS
   return(las)
 }
 
@@ -237,6 +238,7 @@ unnormalize_height = function(las)
   else
     message("No attribute 'Zref' found. Un-normalizisation is impossible.")
 
+  if (las@type > NLAS) las@type <- las@type - NLAS
   return(las)
 }